Migrate data processors in transformation-jvm (#1612)
