+++ streampipes-extensions/streampipes-processors-transformation-jvm/src/main/java/org/apache/streampipes/processors/transformation/jvm/processor/booloperator/timekeeping/BooleanTimekeepingProcessor.java
@@ -18,13 +18,13 @@
 
 package org.apache.streampipes.processors.transformation.jvm.processor.booloperator.timekeeping;
 
+import org.apache.streampipes.commons.exceptions.SpRuntimeException;
 import org.apache.streampipes.model.DataProcessorType;
 import org.apache.streampipes.model.graph.DataProcessorDescription;
-import org.apache.streampipes.model.graph.DataProcessorInvocation;
+import org.apache.streampipes.model.runtime.Event;
 import org.apache.streampipes.model.schema.PropertyScope;
 import org.apache.streampipes.sdk.builder.ProcessingElementBuilder;
 import org.apache.streampipes.sdk.builder.StreamRequirementsBuilder;
-import org.apache.streampipes.sdk.extractor.ProcessingElementParameterExtractor;
 import org.apache.streampipes.sdk.helpers.EpProperties;
 import org.apache.streampipes.sdk.helpers.EpRequirements;
 import org.apache.streampipes.sdk.helpers.Labels;
@@ -32,10 +32,14 @@ import org.apache.streampipes.sdk.helpers.Locales;
 import org.apache.streampipes.sdk.helpers.Options;
 import org.apache.streampipes.sdk.helpers.OutputStrategies;
 import org.apache.streampipes.sdk.utils.Assets;
-import org.apache.streampipes.wrapper.standalone.ConfiguredEventProcessor;
-import org.apache.streampipes.wrapper.standalone.declarer.StandaloneEventProcessingDeclarer;
+import org.apache.streampipes.wrapper.context.EventProcessorRuntimeContext;
+import org.apache.streampipes.wrapper.routing.SpOutputCollector;
+import org.apache.streampipes.wrapper.standalone.ProcessorParams;
+import org.apache.streampipes.wrapper.standalone.StreamPipesDataProcessor;
 
-public class BooleanTimekeepingController extends StandaloneEventProcessingDeclarer<BooleanTimekeepingParameters> {
+import java.util.LinkedList;
+
+public class BooleanTimekeepingProcessor extends StreamPipesDataProcessor {
   // Measures time and returns count
 
   public static final String LEFT_FIELD_ID = "left-field";
@@ -48,6 +52,17 @@ public class BooleanTimekeepingController extends StandaloneEventProcessingDecla
   private static final String MILLISECONDS = "Milliseconds";
   private static final String SECONDS = "Seconds";
   private static final String MINUTES = "Minutes";
+
+  private String leftFieldName;
+  private String rightFieldName;
+
+  private boolean leftFieldLast;
+  private boolean rightFieldLast;
+
+  private LinkedList<Long> allPending;
+  private Long counter;
+
+  private double outputDivisor;
 
   @Override
   public DataProcessorDescription declareModel() {
@@ -80,25 +95,61 @@ public class BooleanTimekeepingController extends StandaloneEventProcessingDecla
   }
 
   @Override
-  public ConfiguredEventProcessor<BooleanTimekeepingParameters> onInvocation(
-      DataProcessorInvocation graph,
-      ProcessingElementParameterExtractor extractor) {
-
-    String leftFieldName = extractor.mappingPropertyValue(LEFT_FIELD_ID);
-    String rightFieldName = extractor.mappingPropertyValue(LEFT_FIELD_ID);
+  public void onInvocation(ProcessorParams parameters,
+                           SpOutputCollector spOutputCollector,
+                           EventProcessorRuntimeContext runtimeContext) throws SpRuntimeException {
+    var extractor = parameters.extractor();
+    leftFieldName = extractor.mappingPropertyValue(LEFT_FIELD_ID);
+    rightFieldName = extractor.mappingPropertyValue(LEFT_FIELD_ID);
     String outputUnit = extractor.selectedSingleValue(OUTPUT_UNIT_ID, String.class);
+    leftFieldLast = false;
+    rightFieldLast = false;
+    allPending = new LinkedList<>();
+    counter = 0L;
 
-    double outputDivisor = 1.0;
+    outputDivisor = 1.0;
     if (outputUnit.equals(SECONDS)) {
       outputDivisor = 1000.0;
     } else if (outputUnit.equals(MINUTES)) {
       outputDivisor = 60000.0;
     }
+  }
 
-    BooleanTimekeepingParameters params =
-        new BooleanTimekeepingParameters(graph, leftFieldName, rightFieldName, outputDivisor);
+  @Override
+  public void onEvent(Event inputEvent, SpOutputCollector collector) throws SpRuntimeException {
+    boolean leftField = inputEvent.getFieldBySelector(leftFieldName).getAsPrimitive().getAsBoolean();
+    boolean rightField = inputEvent.getFieldBySelector(rightFieldName).getAsPrimitive().getAsBoolean();
 
 
-    return new ConfiguredEventProcessor<>(params, BooleanTimekeeping::new);
+    if (!rightFieldLast && rightField) {
+      if (this.allPending.size() > 0) {
+        Long startTime = this.allPending.removeLast();
+
+        Long timeDifference = System.currentTimeMillis() - startTime;
+
+        double result = timeDifference / this.outputDivisor;
+
+        this.counter++;
+
+        if (this.counter == Long.MAX_VALUE) {
+          this.counter = 0L;
+        }
+
+        inputEvent.addField("measured_time", result);
+        inputEvent.addField("counter", this.counter);
+
+        collector.collect(inputEvent);
+      }
+    }
+
+
+    if (!leftFieldLast && leftField) {
+      this.allPending.push(System.currentTimeMillis());
+    }
+  }
+
+  @Override
+  public void onDetach() throws SpRuntimeException {
+
   }
 }
