+++ streampipes-extensions/streampipes-processors-transformation-jvm/src/main/java/org/apache/streampipes/processors/transformation/jvm/processor/csvmetadata/CsvMetadataEnrichmentProcessor.java
@@ -20,12 +20,18 @@ package org.apache.streampipes.processors.transformation.jvm.processor.csvmetada
 
 import org.apache.streampipes.client.StreamPipesClient;
 import org.apache.streampipes.commons.exceptions.SpRuntimeException;
+import org.apache.streampipes.commons.parser.BooleanParser;
+import org.apache.streampipes.commons.parser.FloatParser;
+import org.apache.streampipes.commons.parser.IntegerParser;
+import org.apache.streampipes.commons.parser.PrimitiveTypeParser;
+import org.apache.streampipes.commons.parser.StringParser;
 import org.apache.streampipes.extensions.api.runtime.ResolvesContainerProvidedOptions;
 import org.apache.streampipes.extensions.api.runtime.ResolvesContainerProvidedOutputStrategy;
 import org.apache.streampipes.extensions.management.client.StreamPipesClientResolver;
 import org.apache.streampipes.model.DataProcessorType;
 import org.apache.streampipes.model.graph.DataProcessorDescription;
 import org.apache.streampipes.model.graph.DataProcessorInvocation;
+import org.apache.streampipes.model.runtime.Event;
 import org.apache.streampipes.model.schema.EventProperty;
 import org.apache.streampipes.model.schema.EventSchema;
 import org.apache.streampipes.model.schema.PropertyScope;
@@ -40,31 +46,46 @@ import org.apache.streampipes.sdk.helpers.Filetypes;
 import org.apache.streampipes.sdk.helpers.Labels;
 import org.apache.streampipes.sdk.helpers.Locales;
 import org.apache.streampipes.sdk.helpers.OutputStrategies;
+import org.apache.streampipes.sdk.helpers.Tuple2;
 import org.apache.streampipes.sdk.utils.Assets;
-import org.apache.streampipes.wrapper.standalone.ConfiguredEventProcessor;
-import org.apache.streampipes.wrapper.standalone.declarer.StandaloneEventProcessingDeclarer;
+import org.apache.streampipes.sdk.utils.Datatypes;
+import org.apache.streampipes.wrapper.context.EventProcessorRuntimeContext;
+import org.apache.streampipes.wrapper.routing.SpOutputCollector;
+import org.apache.streampipes.wrapper.standalone.ProcessorParams;
+import org.apache.streampipes.wrapper.standalone.StreamPipesDataProcessor;
 
 import org.apache.commons.csv.CSVParser;
 import org.apache.commons.csv.CSVRecord;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 import static org.apache.streampipes.processors.transformation.jvm.processor.csvmetadata.CsvMetadataEnrichmentUtils.getCsvParser;
 
-public class CsvMetadataEnrichmentController
-    extends StandaloneEventProcessingDeclarer<CsvMetadataEnrichmentParameters>
+public class CsvMetadataEnrichmentProcessor
+    extends StreamPipesDataProcessor
     implements ResolvesContainerProvidedOptions,
     ResolvesContainerProvidedOutputStrategy<DataProcessorInvocation, ProcessingElementParameterExtractor> {
+
+  private static final Logger LOG = LoggerFactory.getLogger(CsvMetadataEnrichmentProcessor.class);
 
   private static final String MAPPING_FIELD_KEY = "mapping-field";
   private static final String CSV_FILE_KEY = "csv-file";
   private static final String FIELDS_TO_APPEND_KEY = "fields-to-append";
   private static final String FIELD_TO_MATCH = "field-to-match";
+
+  private String mappingFieldSelector;
+  private String matchingColumn;
+  private List<Tuple2<String, PrimitiveTypeParser>> columnsToAppend;
+  private Map<String, CSVRecord> columnMap;
 
   @Override
   public DataProcessorDescription declareModel() {
@@ -86,24 +107,6 @@ public class CsvMetadataEnrichmentController
             Arrays.asList(MAPPING_FIELD_KEY, CSV_FILE_KEY, FIELD_TO_MATCH))
         .outputStrategy(OutputStrategies.customTransformation())
         .build();
-  }
-
-  @Override
-  public ConfiguredEventProcessor<CsvMetadataEnrichmentParameters> onInvocation(
-      DataProcessorInvocation graph,
-      ProcessingElementParameterExtractor extractor) {
-    String mappingFieldSelector = extractor.mappingPropertyValue(MAPPING_FIELD_KEY);
-    List<String> fieldsToAppend = extractor.selectedMultiValues(FIELDS_TO_APPEND_KEY, String.class);
-    String lookupField = extractor.selectedSingleValue(FIELD_TO_MATCH, String.class);
-    String fileContents = getFileContents(extractor);
-
-    CsvMetadataEnrichmentParameters params = new CsvMetadataEnrichmentParameters(graph,
-        mappingFieldSelector,
-        fieldsToAppend,
-        lookupField,
-        fileContents);
-
-    return new ConfiguredEventProcessor<>(params, CsvMetadataEnrichment::new);
   }
 
   @Override
@@ -188,5 +191,79 @@ public class CsvMetadataEnrichmentController
 
   private StreamPipesClient getStreamPipesClientInstance() {
     return new StreamPipesClientResolver().makeStreamPipesClientInstance();
+  }
+
+  @Override
+  public void onInvocation(ProcessorParams parameters,
+                           SpOutputCollector spOutputCollector,
+                           EventProcessorRuntimeContext runtimeContext) throws SpRuntimeException {
+    var extractor = parameters.extractor();
+    mappingFieldSelector = extractor.mappingPropertyValue(MAPPING_FIELD_KEY);
+    List<String> fieldsToAppend = extractor.selectedMultiValues(FIELDS_TO_APPEND_KEY, String.class);
+    matchingColumn = extractor.selectedSingleValue(FIELD_TO_MATCH, String.class);
+    String fileContents = getFileContents(extractor);
+
+    try {
+      makeColumnMap(fileContents);
+    } catch (IOException e) {
+      throw new SpRuntimeException(e);
+    }
+    if (this.columnMap.size() > 0) {
+      this.columnsToAppend = fieldsToAppend
+          .stream()
+          .map(c -> makeParser(c, this.columnMap.entrySet().stream().findFirst().get().getValue()))
+          .collect(Collectors.toList());
+    } else {
+      LOG.warn("Could not find any rows, does the CSV file contain data?");
+      this.columnsToAppend = new ArrayList<>();
+    }
+  }
+
+  @Override
+  public void onEvent(Event event, SpOutputCollector collector) throws SpRuntimeException {
+    String lookupValue =
+        event.getFieldBySelector(mappingFieldSelector).getAsPrimitive().getAsString();
+    CSVRecord record = this.columnMap.get(lookupValue);
+    for (Tuple2<String, PrimitiveTypeParser> columnToAppend : columnsToAppend) {
+      event.addField(columnToAppend.k, getRecordValueOrDefault(record, columnToAppend));
+    }
+    collector.collect(event);
+  }
+
+  @Override
+  public void onDetach() throws SpRuntimeException {
+    this.columnMap = new HashMap<>();
+  }
+
+  private Object getRecordValueOrDefault(CSVRecord record, Tuple2<String,
+      PrimitiveTypeParser> columnToAppend) {
+    if (record != null) {
+      return columnToAppend.v.parse(record.get(columnToAppend.k));
+    } else {
+      return columnToAppend.v.parse("0");
+    }
+  }
+
+  private Tuple2<String, PrimitiveTypeParser> makeParser(String columnName,
+                                                         CSVRecord record) {
+    Datatypes columnDatatype = CsvMetadataEnrichmentUtils.getGuessDatatype(columnName, record);
+    if (columnDatatype.equals(Datatypes.Float)) {
+      return new Tuple2<>(columnName, new FloatParser());
+    } else if (columnDatatype.equals(Datatypes.Integer)) {
+      return new Tuple2<>(columnName, new IntegerParser());
+    } else if (columnDatatype.equals(Datatypes.Boolean)) {
+      return new Tuple2<>(columnName, new BooleanParser());
+    } else {
+      return new Tuple2<>(columnName, new StringParser());
+    }
+  }
+
+  private void makeColumnMap(String fileContents) throws IOException {
+    CSVParser parser = CsvMetadataEnrichmentUtils.getCsvParser(fileContents);
+    List<CSVRecord> records = parser.getRecords();
+    this.columnMap = new HashMap<>();
+    for (CSVRecord record : records) {
+      this.columnMap.put(record.get(matchingColumn), record);
+    }
   }
 }
