+++ streampipes-extensions/streampipes-processors-transformation-jvm/src/main/java/org/apache/streampipes/processors/transformation/jvm/processor/value/change/ChangedValueDetectionProcessor.java
@@ -18,27 +18,32 @@
 
 package org.apache.streampipes.processors.transformation.jvm.processor.value.change;
 
+import org.apache.streampipes.commons.exceptions.SpRuntimeException;
 import org.apache.streampipes.model.DataProcessorType;
 import org.apache.streampipes.model.graph.DataProcessorDescription;
-import org.apache.streampipes.model.graph.DataProcessorInvocation;
+import org.apache.streampipes.model.runtime.Event;
 import org.apache.streampipes.model.schema.PropertyScope;
 import org.apache.streampipes.sdk.builder.ProcessingElementBuilder;
 import org.apache.streampipes.sdk.builder.StreamRequirementsBuilder;
-import org.apache.streampipes.sdk.extractor.ProcessingElementParameterExtractor;
 import org.apache.streampipes.sdk.helpers.EpProperties;
 import org.apache.streampipes.sdk.helpers.EpRequirements;
 import org.apache.streampipes.sdk.helpers.Labels;
 import org.apache.streampipes.sdk.helpers.Locales;
 import org.apache.streampipes.sdk.helpers.OutputStrategies;
 import org.apache.streampipes.sdk.utils.Assets;
-import org.apache.streampipes.wrapper.standalone.ConfiguredEventProcessor;
-import org.apache.streampipes.wrapper.standalone.declarer.StandaloneEventProcessingDeclarer;
+import org.apache.streampipes.wrapper.context.EventProcessorRuntimeContext;
+import org.apache.streampipes.wrapper.routing.SpOutputCollector;
+import org.apache.streampipes.wrapper.standalone.ProcessorParams;
+import org.apache.streampipes.wrapper.standalone.StreamPipesDataProcessor;
 
-public class ChangedValueDetectionController
-    extends StandaloneEventProcessingDeclarer<ChangedValueDetectionParameters> {
+public class ChangedValueDetectionProcessor
+    extends StreamPipesDataProcessor {
 
   public static final String COMPARE_FIELD_ID = "compare";
   public static final String CHANGE_FIELD_NAME = "change_detected";
+
+  private String compareParameter;
+  private Object lastObject = null;
 
   //TODO: Change Icon
   @Override
@@ -57,13 +62,29 @@ public class ChangedValueDetectionController
   }
 
   @Override
-  public ConfiguredEventProcessor<ChangedValueDetectionParameters> onInvocation(
-      DataProcessorInvocation graph,
-      ProcessingElementParameterExtractor extractor) {
+  public void onInvocation(ProcessorParams parameters,
+                           SpOutputCollector spOutputCollector,
+                           EventProcessorRuntimeContext runtimeContext) throws SpRuntimeException {
 
-    String compare = extractor.mappingPropertyValue(COMPARE_FIELD_ID);
+    compareParameter = parameters.extractor().mappingPropertyValue(COMPARE_FIELD_ID);
+  }
 
-    ChangedValueDetectionParameters params = new ChangedValueDetectionParameters(graph, compare, CHANGE_FIELD_NAME);
-    return new ConfiguredEventProcessor<>(params, ChangedValueDetection::new);
+  @Override
+  public void onEvent(Event inputEvent, SpOutputCollector collector) throws SpRuntimeException {
+    Object newObject = inputEvent.getFieldBySelector(compareParameter).getRawValue();
+
+    if (newObject != null) {
+      if (!newObject.equals(lastObject)) {
+        lastObject = newObject;
+        //TODO: Better handling of multiple timestamps (if the field "change_detected" is already in the input)?
+        inputEvent.addField(CHANGE_FIELD_NAME, System.currentTimeMillis());
+        collector.collect(inputEvent);
+      }
+    }
+  }
+
+  @Override
+  public void onDetach() throws SpRuntimeException {
+
   }
 }
